+++ NEWS.md
@@ -19,6 +19,7 @@
 
 ## v0.1.1 (2023-06-07)
 
+### Bug fixes
 - Fix Eshell integration
 
 ---