+++ NEWS.md
@@ -1,6 +1,6 @@
 # Urgrep News
 
-## v0.2.0 (in progress)
+## v0.2.0 (2023-08-31)
 
 ### New features
 - Add support for toggling whether to search in hidden files (`M-s h` in the